ng-animation [ref] renamed some files, extracted factories in separate file
