+++ packages/ng/src/app/animations/fade.animation.ts
