+++ packages/ng/src/app/animations/scale.animation.ts
@@ -12,14 +12,14 @@ export const scalingAnimations = (timing: string): AnimationMetadata[] => [
 		animate(
 			timing,
 			style({ transformOrigin: 'left center', transform: 'scaleX(1)', opacity: '1' })
-		)
+		),
 	]),
 	transition('left => void', [ // <--- leaving <---
 		style({ transformOrigin: 'left center', transform: 'scaleX(1)', opacity: '1' }),
 		animate(
 			timing,
 			style({ transformOrigin: 'left center', transform: 'scaleX(0)', opacity: '0' })
-		)
+		),
 	]),
 
 	transition('void => right', [ // <--- entering <---
@@ -27,14 +27,14 @@ export const scalingAnimations = (timing: string): AnimationMetadata[] => [
 		animate(
 			timing,
 			style({ transformOrigin: 'right center', transform: 'scaleX(1)', opacity: '1' })
-		)
+		),
 	]),
 	transition('right => void', [ // ---> leaving --->
 		style({ transformOrigin: 'right center', transform: 'scaleX(1)', opacity: '1' }),
 		animate(
 			timing,
 			style({ transformOrigin: 'right center', transform: 'scaleX(0)', opacity: '0' })
-		)
+		),
 	]),
 
 	transition('void => top', [ // \/ entering
@@ -42,14 +42,14 @@ export const scalingAnimations = (timing: string): AnimationMetadata[] => [
 		animate(
 			timing,
 			style({ transformOrigin: 'center top', transform: 'scaleY(1)', opacity: '1' })
-		)
+		),
 	]),
 	transition('top => void', [ // /\ leaving
 		style({ transformOrigin: 'center top', transform: 'scaleY(1)', opacity: '1' }),
 		animate(
 			timing,
 			style({ transformOrigin: 'center top', transform: 'scaleY(0)', opacity: '0' })
-		)
+		),
 	]),
 
 	transition('void => bottom', [ // /\ entering
@@ -57,14 +57,14 @@ export const scalingAnimations = (timing: string): AnimationMetadata[] => [
 		animate(
 			timing,
 			style({ transformOrigin: 'center bottom', transform: 'scaleY(1)', opacity: '1' })
-		)
+		),
 	]),
 	transition('bottom => void', [ // \/ leaving
 		style({ transformOrigin: 'center bottom', transform: 'scaleY(1)', opacity: '1' }),
 		animate(
 			timing,
 			style({ transformOrigin: 'center bottom', transform: 'scaleY(0)', opacity: '0' })
-		)
+		),
 	]),
 
 	transition('void => center, void => *', [
@@ -72,13 +72,13 @@ export const scalingAnimations = (timing: string): AnimationMetadata[] => [
 		animate(
 			timing,
 			style({ transformOrigin: 'center', transform: 'scale(1)', opacity: '1' })
-		)
+		),
 	]),
 	transition('center => void , * => void', [
 		style({ transformOrigin: 'center', transform: 'scale(1)', opacity: '1' }),
 		animate(
 			timing,
 			style({ transformOrigin: 'center', transform: 'scale(0)', opacity: '0' })
-		)
+		),
 	]),
 ];
