+++ packages/ng/src/app/animations/slide.animation.ts
